Credit Cards Fields Blurred fix
As of now, requests made to Cybersources
fail if using path based MFEs like
tutor is using. This is beccause the
value of the target_origin is `https://apps.bloomed.org.np/payment`,
which the Cybersource doesn't like and sends an error.

This would not be a problem for site like edx because
its payment mfe usl is `https://payment.edx.org/` and the
trailing slash will the taken care of by the `rstrip('/')`
already present in the code.

What this edit does is that it allows people to
set the `payment_mfe_host` in the PAYMENT_PROCESSORS config
inside or cybersource and only falls back...

If that value isn't provided, the fallback is the url of
the payment mfe, which means that sites like edx will
keep working without any edit to be made.